Fix _printf, add stack ret

+++ LngHandler/execution/bin_translating_x86/src/include/stack.h
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <stdint.h>
 
-typedef double Elem_t;
+typedef uint64_t Elem_t;
 
 //---------------------------------------------------------------------------
